Making changes to Uart.c and to SerialMenu.c so uart0 and uart 3 can work at the same time

+++ GlobalFunctions.h
@@ -11,7 +11,7 @@
 #include "fsl_gpio.h"
 #include "DataTypeDefinitions.h"
 
-void delay(uint16);
+void delay(uint32);
